// src/app/modules/lookups/models/static-pages.ts
// Generated by https://quicktype.io

export interface staticPages {
  id:       string;
  value:    string;
  name:     string;
  pageType: PageType;
  pageTypeId: string,
  valueEn:  string
}

export interface PageType {
  id:        string;
  value:     number;
  nameAr:    string;
  nameEn:    string;
  optionSet: OptionSet;
  color:     string;
  isDefault: boolean;
}

export interface OptionSet {
  id:             string;
  name:           string;
  displayNameAr:  string;
  displayNameEN:  string;
  optionSetItems: null[];
}